# src/main/resources/openapi/auth-service-openapi.yaml
openapi: 3.1.0
info:
  title: Authentication Service API - Image Gallery Microservices
  description: |
    Authentication and Token Management Microservice for the Image Gallery application.
    This service handles user authentication, JWT token generation, refresh token management,
    and password storage.
  contact:
    name: cannibal-kektor
    url: https://github.com/cannibal-kektor
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearer-token: []
- internal-service: []
tags:
- name: Authentication API
  description: "API for user authentication, token management, and password storage\
    \ in the Image Gallery system."
paths:
  /api/auth/validate:
    post:
      tags:
      - Authentication API
      summary: Validate access token
      description: Validates the provided JWT access token
      operationId: validate
      requestBody:
        description: Token validation request
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateRequest"
        required: true
      responses:
        "200":
          description: The provided token is valid
        "400":
          $ref: "#/components/responses/ProblemDetailResponse"
        "401":
          $ref: "#/components/responses/ProblemDetailResponse"
        "403":
          $ref: "#/components/responses/ProblemDetailResponse"
        "404":
          $ref: "#/components/responses/ProblemDetailResponse"
        "429":
          $ref: "#/components/responses/ProblemDetailResponse"
        "500":
          $ref: "#/components/responses/ProblemDetailResponse"
        "503":
          $ref: "#/components/responses/ProblemDetailResponse"
      security:
      - internal-service: []
  /api/auth/refresh:
    post:
      tags:
      - Authentication API
      summary: Refresh access token
      description: Generates new access and refresh tokens using a valid refresh token
      operationId: refreshToken
      requestBody:
        description: Refresh token request
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          description: Returning actual tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/ProblemDetailResponse"
        "401":
          $ref: "#/components/responses/ProblemDetailResponse"
        "403":
          $ref: "#/components/responses/ProblemDetailResponse"
        "404":
          $ref: "#/components/responses/ProblemDetailResponse"
        "429":
          $ref: "#/components/responses/ProblemDetailResponse"
        "500":
          $ref: "#/components/responses/ProblemDetailResponse"
        "503":
          $ref: "#/components/responses/ProblemDetailResponse"
  /api/auth/password:
    post:
      tags:
      - Authentication API
      summary: Store user password
      description: Stores hashed user password in the authentication database
      operationId: storePassword
      requestBody:
        description: Password storage request
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StorePasswordRequest"
        required: true
      responses:
        "200":
          description: Password stored successfully
        "400":
          $ref: "#/components/responses/ProblemDetailResponse"
        "401":
          $ref: "#/components/responses/ProblemDetailResponse"
        "403":
          $ref: "#/components/responses/ProblemDetailResponse"
        "404":
          $ref: "#/components/responses/ProblemDetailResponse"
        "429":
          $ref: "#/components/responses/ProblemDetailResponse"
        "500":
          $ref: "#/components/responses/ProblemDetailResponse"
        "503":
          $ref: "#/components/responses/ProblemDetailResponse"
      security:
      - internal-service: []
  /api/auth/logout:
    post:
      tags:
      - Authentication API
      summary: User logout
      description: |
        Invalidates the provided access token and refresh token.
        The access token is blacklisted until its natural expiration.
      operationId: logout
      requestBody:
        description: Logout request with tokens to invalidate
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
        required: true
      responses:
        "200":
          description: "Logout successful, tokens invalidated"
        "400":
          $ref: "#/components/responses/ProblemDetailResponse"
        "401":
          $ref: "#/components/responses/ProblemDetailResponse"
        "403":
          $ref: "#/components/responses/ProblemDetailResponse"
        "404":
          $ref: "#/components/responses/ProblemDetailResponse"
        "429":
          $ref: "#/components/responses/ProblemDetailResponse"
        "500":
          $ref: "#/components/responses/ProblemDetailResponse"
        "503":
          $ref: "#/components/responses/ProblemDetailResponse"
      security:
      - bearer-token: []
  /api/auth/login:
    post:
      tags:
      - Authentication API
      summary: User login
      description: Authenticates user credentials and returns JWT access token and
        refresh token
      operationId: login
      requestBody:
        description: User login credentials
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: Returning actual tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/ProblemDetailResponse"
        "401":
          $ref: "#/components/responses/ProblemDetailResponse"
        "403":
          $ref: "#/components/responses/ProblemDetailResponse"
        "404":
          $ref: "#/components/responses/ProblemDetailResponse"
        "429":
          $ref: "#/components/responses/ProblemDetailResponse"
        "500":
          $ref: "#/components/responses/ProblemDetailResponse"
        "503":
          $ref: "#/components/responses/ProblemDetailResponse"
components:
  schemas:
    ValidateRequest:
      type: object
      description: Request model for JWT token validation
      properties:
        accessToken:
          type: string
          description: JWT access token to validate
          maxLength: 2147483647
          minLength: 30
      required:
      - accessToken
    RefreshRequest:
      type: object
      description: Request model for refreshing JWT tokens
      properties:
        refreshToken:
          type: string
          description: Valid refresh token obtained during login
          maxLength: 36
          minLength: 16
      required:
      - refreshToken
    TokenResponse:
      type: object
      description: Response model containing JWT access token and refresh token
      properties:
        accessToken:
          type: string
          description: JWT access token for API authentication
          readOnly: true
        refreshToken:
          type: string
          description: Refresh token for obtaining new access tokens
          readOnly: true
    StorePasswordRequest:
      type: object
      description: |
        Internal request model for storing hashed user passwords.
        **This DTO is used only for inter-service communication**
        between user-service and authentication-service.
      properties:
        userId:
          type: integer
          format: int64
          description: User id for password storage
          example: 12345
          minimum: 1
        password:
          type: string
          format: password
          description: User password to be hashed and stored
          example: testPassword123
          maxLength: 50
          minLength: 8
      required:
      - password
      - userId
    LogoutRequest:
      type: object
      description: Request model for user logout and token invalidation
      properties:
        accessToken:
          type: string
          description: JWT access token to blacklist
          maxLength: 2147483647
          minLength: 30
        refreshToken:
          type: string
          description: Refresh token to invalidate
          maxLength: 36
          minLength: 16
      required:
      - accessToken
      - refreshToken
    LoginRequest:
      type: object
      description: Request model for user authentication
      properties:
        username:
          type: string
          description: Username for authentication
          example: alex_white
          maxLength: 30
          minLength: 3
        password:
          type: string
          format: password
          description: User password
          example: testPassword123
          maxLength: 50
          minLength: 8
      required:
      - password
      - username
    ProblemDetail:
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
  responses:
    ProblemDetailResponse:
      description: RFC 7807 Error response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  securitySchemes:
    bearer-token:
      type: http
      description: |
        JWT authentication token obtained from the Authentication Service.
        The API Gateway validates the token and forwards user information in headers:
        - X-User-Id: User identifier
        - X-User-Email: User email address
        - X-User-Username: User username
      scheme: bearer
      bearerFormat: JWT
    internal-service:
      type: apiKey
      description: |
        Internal service authentication for inter-service communication.
        Requires the X-System-Internal-Call header with the origin service name.
        Used for service-to-service calls within the microservices ecosystem.
      name: X-System-Internal-Call
      in: header